refact: extrair context menu para ContextMenuActions (actor-sheet.mjs)
